Quick review 1
At least 3 things to note:

- `project.go`: you can return a pointer to a slice of Project in the List, but
you actually declare a slice of pointers to Projects at line 27, and then pass
a pointer to the slice of pointers to the `Do` function.
What you should do is declare `var projects []Project`, pass a pointer to it
to the `Do` function with `&projects` and finally return again a pointer to
this slice at the end of the `List` function. Like I did in the commit.

- `project_test.go`: you wrote `want := []*Project{{Id: Int(1)}, {Id: Int(2)}}`.
But you are directly declaring a simple `[]Project` type, not a pointer to it.
The type is actually just a slice of Project.

- `project_test.go`: When you write a test, you don't actually test the value
and print an error yourself. You should use the `assert` lib. Also note that
the last asset that I wrote tests two pointers, but, the documentation states
that "Pointer variable equality is determined based on the equality of the
referenced values (as opposed to the memory addresses)", so it's fine.
Read here : https://pkg.go.dev/github.com/stretchr/testify/assert#Equal

The last error you got is probably because your request doesn't actually
returns the body you expect, but probably a simple number instead, might
be an error code. You should check the response's status code before
attempting to unmarshal the body. Maybe have a look at line 66 of `client.go`
of the project I shared with you. It's the `sendRequest` function.